Created the last unit tests for configuration application for the URLSession and CRT
